requirements.txt
streamlit
pandas
numpy
plotly
openai
python-dotenv
openpyxl